+++ src/types/index.ts
@@ -1,3 +1,0 @@
-export * from './color'
-export * from './theme'
-export * from './vsc-theme'